# docs/swagger.yaml
basePath: /
definitions:
  core.Customer:
    properties:
      customer_uid:
        type: string
      email:
        type: string
    type: object
  core.Delivery:
    properties:
      address:
        type: string
      city:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      region:
        type: string
      zip:
        type: string
    type: object
  core.Item:
    properties:
      brand:
        type: string
      chrt_id:
        type: integer
      name:
        type: string
      nm_id:
        type: integer
      price:
        type: number
      rid:
        type: string
      sale:
        type: integer
      size:
        type: string
      status:
        type: integer
      total_price:
        type: number
      track_number:
        type: string
    type: object
  core.Order:
    properties:
      customer_id:
        type: string
      date_created:
        type: string
      delivery:
        $ref: '#/definitions/core.Delivery'
      delivery_service:
        type: string
      entry:
        type: string
      internal_signature:
        type: string
      items:
        items:
          $ref: '#/definitions/core.Item'
        type: array
      locale:
        type: string
      oof_shard:
        type: string
      order_uid:
        type: string
      payment:
        $ref: '#/definitions/core.Payment'
      shardkey:
        type: string
      sm_id:
        type: integer
      track_number:
        type: string
    type: object
  core.Payment:
    properties:
      amount:
        type: number
      bank:
        type: string
      currency:
        type: string
      custom_fee:
        type: number
      delivery_cost:
        type: number
      goods_total:
        type: number
      payment_dt:
        type: integer
      provider:
        type: string
      request_id:
        type: string
      transaction:
        type: string
    type: object
host: localhost:8000
info:
  contact: {}
  description: API Server for WB-level-0 Application
  title: WB API
  version: "1.0"
paths:
  /api/v1/customer:
    post:
      consumes:
      - application/json
      parameters:
      - description: New customer
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.Customer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Customer'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create Customer
      tags:
      - customer
  /api/v1/customers/{uid}/orders:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/core.Order'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get customer orders
      tags:
      - order
  /api/v1/delivery:
    post:
      consumes:
      - application/json
      parameters:
      - description: New delivery
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.Delivery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Delivery'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create Delivery
      tags:
      - delivery
  /api/v1/item:
    post:
      consumes:
      - application/json
      parameters:
      - description: New item
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.Item'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Item'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create Item
      tags:
      - item
  /api/v1/order:
    post:
      consumes:
      - application/json
      parameters:
      - description: New order
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.Order'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Order'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create Order
      tags:
      - order
  /api/v1/payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: New payment
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/core.Payment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/core.Payment'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create Payment
      tags:
      - payment
swagger: "2.0"